MIDI: Update descriptions of Timidity configuration
Simplify overview so the instructions are aimed at someone who wants
to make a game and use MIDI.

The correct location of the configuration file appears to be the game
directory and not the engine directory.

Remove references to id Software and Doom. It isn't clear that the
patches have anything to do with Doom, the description implies the
patches are taken from a driver.

Add legal warning because it isn't clear that distributing the linked
patches is allowed (especially if there are from Doom or a from a
propriatary driver).

List additional configuration path for Linux and macOS (the
correctness of these is likely determined by the Timidity build
process on those platforms).

Document environment variable TIMIDITY_CFG.

Remove bad Timidity configuration examples for dir directive ("etc/"
is a relative path).

Document Timidity configuration token $basedir.